Prepare for clustering with improved name support
Prepare for the setup of the clustering by adding configuration support
for setting the cluster name. This will be used for logging, providing a
differentiator, and for metrics. Additionally, re-rename the endpoint
label to component in order to avoid conflicts with pre-defined
Prometheus labels, which is causing endpoint to be renamed to
exported_endpoint, just like with service.