Update dashboard to use the Pod IP as listening address
Update the dashboard application to use the Pod IP as the listening
address, which should help improve the data in the log files by allowing
that address to be used instead of the generic 0.0.0.0.

This is being done through the environment variable setting overrides
using viper as it's easier than setting it via the command-line.